Restructured and commented implementation.
