+++ CMakeLists.txt
@@ -3,4 +3,4 @@ project(Project)
 
 set(CMAKE_CXX_STANDARD 17)
 
-add_executable(Project main.cpp Vector.hpp constants.hpp Tree.hpp)
+add_executable(Project main.cpp Vector.hpp constants.hpp Tree.hpp Particle.hpp Cell.hpp)
